# Optimize cell detection