--- a/cellfinder/core/detect/detect.py
+++ b/cellfinder/core/detect/detect.py
@@ -163,7 +163,7 @@ def main(
     # processing so for n_ball_procs 2d filtering threads, ball_z_size will
     # typically be in us while the others stall waiting for 3d processing
     # so we can use those for other things, such as numba threading
-    set_num_threads(max(n_ball_procs - ball_z_size, 1))
+    set_num_threads(max(n_ball_procs - int(ball_z_size), 1))
 
     start_time = datetime.now()