--- a/cellfinder/core/detect/filters/volume/ball_filter.py
+++ b/cellfinder/core/detect/filters/volume/ball_filter.py
@@ -254,7 +254,6 @@ class BallFilter:
             )
 
 
-
 @njit(cache=True)
 def _cube_overlaps(
     volume: np.ndarray,