… .pre-commit-config.yaml
… cellfinder/core/detect/filters/volume/structure_splitting.py
… cellfinder/core/detect/filters/volume/volume_filter.py
… tests/core/test_unit/test_detect/test_filters/test_volume_filters/test_structure_detection.py